feat: Support times_per_day habit frequency
Adds a new `times_per_day` frequency option to the `HabitFrequency` type. This change also updates the `Habit` interface and the `Rituals` component to accommodate tracking habit completion by count per day, in addition to existing frequency types. This enables users to set and track habits that need to be performed a specific number of times each day.