refactor: Remove unused Tooltip component
The `Tooltip` component is no longer used in the `Layout` or `ProgressStats` components and has been removed.

The `ProgressStats` component was refactored to dynamically generate tooltip classes based on progress values, removing the need for a separate generic tooltip component.

The `Layout` component also previously utilized the `Tooltip` component, but this usage has also been removed.